+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added options to restore notes from trash
   - With the swipe actions of a note
   - With the "Restore from trash" button in the multi select options
+- Added option to create a logfile for sync 
 - Added experimental option to enable virtual folder tags
 
 ## 0.1.7